+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fixed a bug where nilable and blankable Spec conformers would return `INVALID` even
+  if their validation passed (#87)
 
 ## [v0.3.post1]
 Fixed `README.rst` reference and content-type in `setup.py` for PyPI description.